test(emitter): split ES5 class downlevel tests
